backend/python-service/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
PyMuPDF==1.19.6
Pillow==9.3.0
flask-cors==3.0.10
